Add local storage function file
