+++ src/index.js
@@ -1,9 +1,11 @@
 import { join } from 'lodash';
 import './style.css';
-import { check } from './modules/check';
-// import { storage } from './modules/storage';
+// import { check } from './modules/check';
+import { storage } from './modules/storage';
 
-export const tasks = [
+let tasks = [];
+if (localStorage.length === 0) {
+ const tasks = [
   {
     description: 'Wash dishes',
     completed: false,
@@ -20,9 +22,9 @@ export const tasks = [
     index: 3,
   },
 ];
-
-localStorage.setItem('tasks', JSON.stringify(tasks));
-
+} else {
+  tasks = storage.getLocal();
+}
 const displayTasks = () => {
   const listTasks = document.querySelector('.listTasks');
   for (let i = 0; i < tasks.length; i += 1) {
@@ -40,8 +42,31 @@ const displayTasks = () => {
   }
 };
 
+const check = () => {
+  const checkboxed = document.getElementsByClassName('checkboxed');
+  for (let j = 0; j < checkboxed.length; j++) {
+    checkboxed[j].addEventListener('change', () => {
+      if(checkboxed[j].checked) {
+        checkboxed[j].nextElementSibling.style.textDecoration = 'line-through';
+        tasks[j].completed = true;
+        console.log(tasks[j].completed);
+        localStorage.setItem('tasks', JSON.stringify(tasks));
+      } else {
+        checkboxed[j].nextElementSibling.style.textDecoration = 'none';
+        tasks[j].completed = false;
+        console.log(tasks[j].completed);
+        localStorage.setItem('tasks', JSON.stringify(tasks));
+      }
+    });
+  }
+}
+
 displayTasks();
 check();
+
+
+
 if (localStorage.length !==0) {
- tasks = JSON.parse(localStorage.getItem('Tasks'));
+ const storedTasks = JSON.parse(localStorage.getItem('Tasks'));
+ displayTasks(storedTasks);
 }
